Modified: given 3Floats of maximum

+++ src/com/bridgelabz/generics/MaxValue.java
@@ -1,7 +1,7 @@
 package com.bridgelabz.generics;
 
 public class MaxValue {
-    void maxValue(Integer val1,Integer val2,Integer val3){
+    void maxValue(Float val1,Float val2,Float val3){
         if((val1.compareTo(val2))>=0 && (val1.compareTo(val3))>=0)
             System.out.println(val1+ " is the largest");
